Refaktorer VedtakBegrunnelseSpesifikasjon
